Fix: double->float 변경

See merge request s08-bigdata-recom-sub2/S08P22A804!227

+++ backend/incense/src/main/java/com/suyang/incense/api/response/analysis/NoteGraphDto.java
@@ -19,7 +19,7 @@ public class NoteGraphDto {
     public static class NoteInfo{
 
         private String word;
-        private double weight;
+        private float weight;
 
     }
 }
